fix template tiddler reference
This reference was hard-coded and I changed the name when I moved it to
a plugin.

In the future, we'll want to move the rendering template out of the
macro. I just put it in there as a hack while I didn't understand how
rendering worked and wanted to move on for the time being. Saq
kindly helped me figure out what was going on on the TW forums, so I
should now be able to do it properly.