store deck and tag values in TwNote
These values are not yet used for anything, but existing functionality
should continue to work normally and this sets the stage for actually
redirecting notes.

The config tiddlers work just like $:/config/FileSystemPaths on Node:
you put in a series of newline-separated filters that are matched to
each tiddler. For Tags, all lines that yield output are treated as tags.
For Deck, the first line that yields output is treated as the target
deck.

We may need to do some munging on the result since, e.g., Anki tag names
cannot contain spaces.