make TR recognize when tags are unchanged
canonify() seems to be returning empty strings along with the actual
canonified tags? It was recently rewritten in Rust, so perhaps a bug was
introduced then, but I don't know enough about Rust collections to spot
if there's an issue, and I don't feel like debugging Anki core right
now, so we'll just remove them if they show up. The empties get removed
when you actually flush to the database it seems, but having them there
was confusing our equality checks and making TR think it had to update
every existing note every time because the tags had "changed" to include
empty strings.

I also added !is[blank] to the filters. At first I thought this was the
source of the empty strings; now I am still leaving it in even though
it's not because skipping multiple newlines without confusing anything
is a good thing to do!