track wiki as a field on each note
This does *not* change the uniqueness constraint on IDs. That means that
notes can be moved between wikis and wikis can be renamed losslessly,
which is much better.

If we're still worried about ID collisions, adding an option to
prefix/suffix all IDs on the TiddlyWiki side could help. I'm not worried
for now, but if TiddlyRemember against all odds ends up taking over the
world and people start trying to share notes, we could conceivably end
up with collisions. Millisecond timestamps aren't GUIDs or anything.