allow changing from cloze to question type or v.v.
Obviously, this can't be completely lossless -- the scheduling
information may not apply exactly, depending on what kind of
change was made. Cloze cards after the first are lost when converting
to question format, and new ones may be generated when converting
to cloze format. Nonetheless, it's nice to have the option to convert
without causing a huge hullabaloo.