check to make sure user hasn't modified fields
I have encountered at least one user who accidentally changed the LPCG
note type's fields and blew something up. The errors provided may not be
helpful to someone who doesn't know Anki (I don't believe it's possible
to explain what went wrong and how to fix it in the space of an error
message!), but issuing an error message is still way better than
accidentally trashing notes in some way, which might be the alternative.

I also moved ensure_note_types() to the start of the sync function, both
for consistency/simplicity and to guard against the possibility that a
user deletes the note type between starting Anki and trying to sync.