fix tiddler titles with ?, :, etc., breaking sync on Windows
TiddlyWiki replaces Windows-invalid characters with _ when saving
tiddlers to a folder wiki, but not (as I had expected) when generating
HTML files. URL-encoding the tiddler names is a quick way to make them
valid filenames on all operating systems; since we won't be doing
anything with these other than reading through them programmatically
and then deleting them, the fact that it produces ugly names is of no
consequence.