minimum working Anki version is 2.1.24, not 2.1.20
The form `find_notes` was only introduced in 2.1.24. Rather than editing
the code to make it backwards-compatible, I'm just going to increase the
minimum version...the need to update Anki by a couple of patch versions
should not prevent anybody from being able to use TiddlyRemember, so
it's not worth the trouble.

Fixes #8.