allow the inclusion of hard references in remember* calls
A "hard reference" is a tiddler reference whose name is stated in
the macro call, in contrast to a "soft reference", which is obtained
from the tiddler that the macro call is rendered within. A soft
reference will be used if no hard reference is specified with the use of
the optional "reference" parameter to remembercz.

The reference parameter probably really should come before the mode
parameter of remembercz for symmetry, but it's definitely not worth
breaking backwards compatibility!

Currently, this yields no changes in behavior whatsoever -- the div is
not displayed in the browser and the Anki add-in doesn't pick it up
either.