fix note deletion not working in Anki 2.1.28 and up
2.1.28 moved the note-deletion code to the Rust backend (commit
4a69b55a90), and it appears that in this process, the series of note IDs
to delete changed to supporting only a Sequence rather than an Iterable.
For some reason passing an Iterable didn't yield an error, it just
didn't do anything at all. In any case, dropping the IDs into a list
before sending them through fixes it up.

Also changed the method name to remove_notes, as the old remNotes name
was deprecated as part of this change.