test import into Anki
This proved easier than I expected -- the only aqt-related object I was
using in the non-UI code was mw.col, and I was able to get around that
by passing a /col/ parameter into everywhere I used that. It's easy to
open a collection outside the GUI, so that makes things easy.

Line coverage is over 80% just like this.

The one irritant is that the minimum Anki version is now 2.1.28, because
canonifying tags required a mw instance
prior to that version. Happily, this has been fixed in .28+.