import HTML into Anki content fields
This currently takes a rather poor approach to links, just wiping them
out entirely, because otherwise the cards are full of links that don't
work. What we should do is figure out if they're internal or external,
and try to use the permalink to change to external if they're internal.