handle links in imported notes
External links have all their target, class, etc., properties removed to
make them into "normal" links.

Internal links are simply zapped. I originaly wanted to make them into
permalinks, but this is harder than it sounds because the permalink
field isn't known or filled in until quite a bit later in the sync
process. Of course we could drop in a placeholder and delay filling it
in, but I'm not convinced this is worthwhile, especially since in some
cases the links will be more annoying than helpful and we'd probably
want to put it behind an option anyway. If you do want to click the
link, you can always just go to the permalink and click it from there.

I'm guessing that nobody will even ask for this, but if they do we can
always add it then.