tweak margin and line-height to improve display of ID
Previously, the paragraph under the ID display would wrap prior to
reaching the ID in brackets, even though it looked like it could *just*
squeeze in under the ID. This puts a margin on the question element so
enough space is cleared out beneath the float.

Ideally, the ID would float *up* into the space to the right of the
answer, if the last line of the answer were short enough. That is beyond
my CSS skills at this point.

Fixes #21.