offer useful error message if a tiddler name is too long
Addresses #14. Since tiddler names are used as filenames, they can
cause OS path length limitations to be exceeded, especially since they
get URL-encoded and thus get a little bit longer when punctuation
or non-ASCII characters are in the mix.

A more robust possibility would be to use some kind of generated unique
key for the filename that is quite short (e.g., a hash of the title),
then embed the tiddler title in the template itself as an HTML element
that can be parsed by TiddlyRemember. However, I'm not actually aware
of a filter that can do this in TR for the time being, and I've only
seen this surface once so far (and that from a user who admitted it was
a "pathologically long" tiddler name). So for now we will give an error.