allow checking for incompatible TW plugin version
Fixes #30.

Notes:

- Currently "" is an acceptable TW plugin version, since no version
number was provided by the previous version, which is compatible.

- This adds two more places to update the version. I added them to
RELEASING.md, but in the future it might be worth automating version
bumps and/or working on transclusions between the docs and plugin!

- Right now the version is verified on every tiddler. It should really
be per wiki, but that would require some ugly flags or a separate
version-verifying step using a separate template, and it doesn't
seem like this is causing much performance degradation, so we'll leave
it this way for now.