support syncing KaTeX from TiddlyWiki
With the official plugin "tiddlywiki/katex", you can include equations
in your TiddlyWiki. However, this doesn't come across neatly in Anki.
This change searches for KaTeX spans and trims out the cruft that
actually displays the markup in a browser, then extracts the raw markup.
Then it wraps that markup in \( )\ or \[ \], the syntax for the MathJax
implementation that Anki has integration since 2.1.

There is potential for some kind of incompatibility between the KaTeX
and MathJax implementations, but I'm not aware of any at this time, and
at any rate there's not a whole lot to be done about it.

There are native MathJax plugins for TiddlyWiki also, but at the time of
this writing I was unable to get them to work in TiddlyWiki at all, so
trying to sync them seems premature!

Fixes #28.