request verbose output when rendering tiddlers
I wanted to indicate exactly which tiddler broke it, but it turns out
that TiddlyWiki doesn't output this information before crashing (boo),
so this is the best we can do (one can at least figure out which tiddler
comes next alphabetically, given this output). I submitted an issue to
core TiddlyWiki requesting this be improved:

https://github.com/Jermolene/TiddlyWiki5/issues/6067

Temporary improvement for #44.