allow importing scheduling information
Scheduling info is stored in a SchedulingInfo object inside notes, and
this information is applied to the collection when notes are new. Anki's
copy takes precedence if it exists.

The TR macros provide the interface to scheduling as a single
stringly-typed variable called "sched" in which we can add the necessary
information. That seems like a dirty hack, but I suspect this is better
than adding *four* additional parameters which most people will never
use.