allow escaping braces in cloze text
\{ or \} turns the brace into a standard brace. Backslashes not
preceding a brace are treated as literal. \\{ produces \{. A backslash
prior to an actual cloze deletion is not possible (this isn't ideal but
doesn't seem like a big enough imposition to justify rewriting the
parser to use something more complicated than a regex with a
lookbehind).

In addition, since we can no longer abuse Python's format string code to
create the final Anki syntax because that uses single braces as
placeholders and some of the single braces can be literal now, we switch
to using a second regex replace on the placeholder syntax ©<0>© (I
figure only a crazy person would use those characters together). This
means that we no longer have to throw an exception and abort the sync if
braces are mismatched. The card will still likely come out malformed (of
course), but the problem should be obvious when the user sees it, and he
can then go correct it.

Resolves #43.