add "class" attribute to remember* macros
This allows users to conveniently apply arbitrary styling to either
individual TR notes or all TR notes in the wiki (via a config tiddler
that contains classes applied to every call). In this way, it's the
opposite of changing the card templates in Anki -- you can style only
the wiki version and not the Anki version.

Partially completes #46 (covers removing the text "cloze", or changing
it to something via a CSS text replacement, as well as hiding the
self-id of inline cloze deletions or anything else).