implement media syncing for images
We extract the 'src' attribute from any img tags and retrieve the URL
from it, then add it to Anki's media folder if it doesn't already exist.

Syncs happen on both adds and updates, and all files must be downloaded
in order to hash them and determine if they need to be synced in. This
could potentially be sped up through some sort of cache if people start
complaining about long sync times. Or the people complaining could just
actually download the images and put them on their hard drives.

Valid links external to the wiki, _canonical_uri tiddlers with absolute
paths, and embedded tiddlers all work great as image targets. We can
probably make _canonical_uri tiddlers with relative paths work too by
playing with the current directory; that's an opportunity for
improvement.

If some images can't be retrieved, a warning dialog appears and the user
has the opportunity to abort the sync if she thinks the problem is
transient or can be quickly fixed. Any images that can't be retrieved
have the <img> tag in TiddlyWiki passed through to Anki as-is, rather
than having the 'src' attribute replaced with the newly synced image in
Anki's media folder.

Old media are not deleted on deletes; the media check functionality in
Anki can be used to clear them up if needed (this is how other parts of
Anki work too, I'm not just being lazy).