make media import of _canonical_uri tiddlers work
Relative paths in this field work for file or URL wikis. (They don't
work for folder wikis in TiddlyWiki either since there's a web server
essentially running in a chroot.)

Includes refactor: add metadata about a wiki to a 'wiki' object in
twimport.py that gets passed down to the places that the wiki_name
string used to be passed. We needed the source URL and the type in
extract_media for this change, and this way we can add as much
additional metadata as we need without adding even more arguments that
have to be thrown around.