support syncing and exporting audio files
With the infrastructure for images in place, this was a piece of cake.

We might want to support some other ways of including audio besides
HTML5 <audio> tags, but I'll wait until someone actually wants that and
explains what method they want and why they want it before going to any
more trouble.