don't display \{ escapes in TiddlyWiki
KaTeX displays incorrectly if the backslashes aren't pulled out prior to
display, and it's kind of ugly anyway. Escapes can stay in the realm of
the source text.

We add a new "cloze-display" span inside clozes, preserving "cloze-text"
to be used by the parser.

A future possibility would be to colorize the brace pairs that are
active/unescaped in cloze-display.