allow users to easily change snippets
`remembercz` and `rememberq`'s default snippets can be edited directly,
and instructions are provided in the documentation for duplicating the
pattern to create other buttons entirely.

%NOTE_ID% and %EDITING_TIDDLER% placeholders are implemented
specifically for this because the contents of the snippets are macros
which are used as raw text, not evaluated before use. This makes it
impractical to use the standard <<macro>> and {{field}} transclusion
syntax to include variable parts in the macros.

I hope this will resolve #46.