choose qt5 or qt6 UI as needed for current runtime
The approach of generating pyuic5 and pyuic6 code for the same *.ui
files means that we won't be able to use any new features in Qt 6 and we
should make sure we edit the files with the Qt 5 Designer, but this
appears to work perfectly fine for the moment, so there's no reason to
do dual maintenance before we need anything else. I imagine we'll be off
Qt 5 for supported versions in a few months anyway.

Confirmed that this works on both the current and beta (with Qt 6)
versions of Anki.