fix password not being saved when config file created on old version
If the 'password' field wasn't in the meta.json, we wouldn't try to save
the contents of the password field, since we were iterating over each
key in the current schema and matching it up with the appropriate field.
To fix this, just add it with a blank value if it doesn't already exist
before trying to do the save routine.

Fixes #61.