update env for latest Anki version
Note that Anki actually is showing PyQt 6.6.1 rather than 6.8.1, but
when I tried using 6.6.1 it failed to link against my system libraries
and collapsed in a puddle of green goo. I hope using 6.8.1 is safe – I
don't think we really test anything deeply related to PyQt anyway, so I
imagine it'll be fine.